requirements.txt
streamlit
langchain
langchain-community
langchain-mistralai
chromadb
pypdf
python-dotenv
tiktoken
altair<5
# Lexis AI Book Assistant - Production Requirements
# Python 3.10+

# Core Framework
streamlit>=1.28.0,<2.0.0
streamlit-chat>=0.1.1

# LangChain & AI
langchain>=0.1.9,<0.2.0
langchain-community>=0.0.25,<1.0.0
langchain-core>=0.1.30,<0.2.0
langchain-mistralai>=0.1.0,<1.0.0

# Vector Database & Embeddings
chromadb>=0.4.0,<0.5.0


# PDF Processing
pypdf>=3.17.0,<4.0.0
pdfplumber>=0.9.0,<1.0.0

# Utilities
python-dotenv>=1.0.0,<2.0.0
pydantic>=1.10.0,<2.0.0


# Optional: For advanced features
requests>=2.31.0,<3.0.0
cachetools>=5.3.0,<6.0.0

# Development (optional)
# pytest>=7.4.0
# black>=23.0.0
# flake8>=6.0.0
# mypy>=1.5.0
